+++ test/src/shared/interfaces/ICardBack.ts
@@ -1,0 +1,14 @@
+export interface ICardBack {
+    cardName: string;
+    color: string;
+    skillDescription: string;
+    power: string;
+    character: string;
+    specialTrait: string;
+    skillKeywords: string[];
+    relatedCharacters: string[];
+    relatedSpecialTraits: string[];
+    relatedCardNames: string[];
+    era: string;
+    cardImageUrl: string;
+}
